backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
bcrypt==4.0.1
python-dotenv==1.0.0
psutil==5.9.5
Werkzeug==2.3.7
gunicorn==21.2.0
